Catch schema parsing errors in DataPlatformIterableSource (#4305)
Prior to this change, the data platform iterable source would fail initialization when it encountered a schema it could not parse. This would prevent the user from looking at any data if only one topic had a schema problem.

This change catches schema parsing errors and adds them to initialization problems. This allows the user to continue using other topics which did not have errors.