3D Beta: add onboarding tooltip on settings button (#4358)
**User-Facing Changes**
None (3d beta, self-explanatory feature)

**Description**
Adds a tooltip to help users discover topic settings, which are in a
significantly different location from the old panel.

The tooltip is shown on any new 3D panel, until the first time the user
clicks the settings button. The clicked state is stored in the user
profile (in localStorage).

Also updates thumbnail image and description for 3D beta panel

Closes #4354

Co-authored-by: Andrew Carr <andrew@foxglove.dev>
Co-authored-by: Adrian Macneil <adrian@foxglove.dev>