Move settings actions after error icon (#4380)
The settings actions are represented as a context menu icon. The typical
pattern for context menu icons is to appear at the start or end of the
component (or line) as the vertical dot icons are representative of a
"more" action.

This change moves the actions icon after the error icon in the
NodeEditor.

The error indicator was placed after the actions in
https://github.com/foxglove/studio/pull/3384/ which correctly moved the
actions after the visibility icon but also placed them before the error
icon rather than at the end.

Fixes: #4268


<!-- link relevant github issues -->
<!-- add `docs` label if this PR requires documentation updates -->