Add support for reading unindexed mcap files (#4391)
This brings back the same level of support for unindexed mcap files that
we had in the previous player architecture. The unindexed mcap file is
fully read into memory. For now this is the only way to process an mcap
file with no index or summary section because the IIterableSource can
only provide topic information and datatypes during initialization not
during message reading.

Future enhancements could include:
- Reading the summary section if it contains schemas and channels only
thus avoiding the need to read the message data into memory.
- Adding a way for IIterableSource to provide updates to topics and
datatypes as messages are read.

Fixes: #4389