Fix DataPlatformIterableSource partial consumption (#4449)
When streaming from data platform using the "partial" consumption type,
the iterable source would incorrectly reset the next window. This
happened when trying to use coverage information to skip gaps but
failing to detect that the window was already in a valid coverage zone.

Regressed in: https://github.com/foxglove/studio/pull/4398