Hides Clear button when color is cleared or when flag is set to true (#4486)
**User-Facing Changes**
The Clear button is now hidden whenever there is no color defined in the
color picker or when the panel author chooses to hide it.

**Description**
This adds a new flag for color pickers, `hideClearButton`, which hides
the clear button when set to true by the panel author.

On top of covering the case where the color value is cleared, this
allows flexibility for different use-cases, such as if different color
values are being generated dynamically or if the author conditionally
wants or doesn't want users to change the color picker (such as hiding
when a user-defined default value is chosen).

Fixes #4122