Add datatype to MessageEvent (#4554)
Prior to this change each MessageEvent only had a topic name. If a panel
author wanted to determine the type of message within the MessageEvent
they would need to keep a mapping of topic -> datatype.

This change adds the `datatype` field to the MessageEvent. This field
contains the datatype of the message within the message event.

Fixes: https://github.com/foxglove/studio/issues/4561