Support foxglove.Grid in 3D Panel(#4544)
**User-Facing Changes**
- enable visualization of Foxglove.Grid topic data types in the 3D panel

**Description**
 - Existing issues with this implementation: 
- ~~converted "/map" OccupancyGrid message takes much more time to
update the exture because of the CPU color converters. Do we want this
implemented with shaders for this initial pass, sounds like it's a TODO
for right now for Pointclouds.~~ Should be done in shaders in future --
similar to point clouds
- ~~There's some pretty bad z-fighting without `logarithmicDepthbuffer`
enabled -- not sure what we want the solution to be here. I don't know
why it's acting differently from the OccupancyGrid~~ This might be added
as a setting in the future
- Adapted a lot of code from the `OccupancyGrid` and
`PointCloudsAndLaserScans` renderables
- tried to genericize and move ColorMode/field utility functions. I did
not make the pointclouds use the settings nodes because I didn't want to
break anything in this already large PR, so I'll do it in a followup.
 - created `Grid`s stories


<!-- link relevant github issues -->
Fixes #4465
<!-- add `docs` label if this PR requires documentation updates -->

Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>