[3D] URDF joint viewer, manually modify joint positions (#4573)
This updates the `Urdfs` SceneExtension to display a list of all joints
for a URDF and all parsed information for each
joint. A "manual" field for each joint controls the angle or position by
modifying the offsetPosition or
offsetEulerDegrees for the relevant frame.

Support for JointState is partially added, but the subscription won't
work
until we have a better way for scene extensions to signal whether a
particular subscription should be on or off (in the
case of Urdfs, we want to enable the subscription when there is at least
one loaded URDF).