Fix Gauge/Indicator skipping messages on high-frequency topics (#4671)
**User-Facing Changes**
Fixed a bug in the Gauge and Indicator panels where messages on
high-frequency topics would sometimes be skipped when using a message
path filter.

**Description**
These panels only processed the _last_ message on each `currentFrame`.
That's incorrect, because if the message path contains a filter, one of
the intermediate messages in the frame might match the filter even if
the last one does not. To fix the issue, we process all messages in the
currentFrame on the requested topic. Also added stories with fixture
data containing multiple messages per frame to prevent regressions.