3D: fix square point shape for point clouds (#4683)
**User-Facing Changes**
[3D] Fixed "square" point shape setting for point clouds.

**Description**
Fixes #4679.
This looks like it has been broken since #3420 (5 months)?! 😧 It was
working for LaserScans, but not PointClouds.

The error was `'sRGBToLinear': no matching overloaded function found` i.e. a missing function definition. The function definition is in the `FS_SRGB_TO_LINEAR` constant and was only being added in the circle code path.