Improve plot and state transition panel zoom state handling (#4606)
Change ChartJSManager to respect and not change existing, user
manipulated scales except in two cases:
1. When an explicit min/max are specified in settings those take
precedence.
2. When we pass an explicit `isBoundsReset` flag in the update message.

The second case is necessary because a min/max of `undefined` is
currently interpreted as an intention to let the chart determine its own
bounds based on the dataset but this has the side effect of clearing any
change in the scale the user has initiated directly, so we want to avoid
this unless the user has clicked the "reset view" button, which we
currently interpret as a temporary absence of global bounds.

Fixes #4351
Fixes #4684