Fix variable sidebar value updates (#4724)
Fixes logic around updating externally changed values in the variable
sidebar.

There is an `isEditing` flag we use to determine if we should apply new
values from changes not made directly in the sidebar but once this is
set to true there is no way for it to become false again. This uses the
editor's `onBlur` event to set it back to false so we can accept
external values again.