Fix: the issue where the map does not display when opening the map topic in some cases.
First of all, some topics do not have data in every frame, such as maps. Map data only exists in the first frame since sending a package of unchanging data for every frame would be unnecessarily large. The existing logic checks previous frames for missing data in subsequent plays until the data is found, reusing cached data from previous frames.

However, in many cases, the previous frame data for certain topics is not cached, such as when seeking back and forth or opening a topic in the middle of data. Previously, we sent a getStream request (snapshot request) with a start time equal to the end time, which returned complete data and allowed the backend to find the previous frame data. The current issue is that toggling the topic does not send a snapshot request, resulting in incomplete data. Inserting a snapshot request fixed this issue, but during playback, getStreams requests are already sent to request the next segment of data, making additional snapshot requests redundant.

We decided to add a parameter to the getStreams to inform the backend whether a complete frame of data is needed.